Improve overlap performance (#111)
This PR adds the option to calculate the overlap only as a triangular
matrix. This reduces the number of pairs over which we have to iterate
in order to write the overlap matrices obtained from the MMD algorithm
in the correct position of the full overlap matrix. As this loop appears
to be very costly, a significant amount of time can be saved.

---------

Co-authored-by: hoelzerC <51398942+hoelzerC@users.noreply.github.com>