Collect classical contributions similar to interactions (#119)
This should allow for more flexibility for the classical contribution.
They can now be passed to the `Calculator` in the same way the
self-consistent contributions (`Interaction`s) are passed. They are then
collected with the typical contributions in a `ClassicalList`, which
internally iterates over all contributions carrying out `get_cache`,
`get_energy`, and/or `get_gradient`.